+++ R/leaflet.esri-package.r
@@ -1,4 +1,9 @@
-#' leaflet.esri.
+#' leaflet.esri: 'ESRI' bindings for the 'leaflet' Package.
+#'
+#' An add-on package to the 'leaflet' package, which provides bindings for 'ESRI' services.
+#' This package allows a user to add 'ESRI' provided services such as
+#' 'MapService', 'ImageMapService', 'TiledMapService' etc. to a 'leaflet' map.
+#'
 #' @importFrom magrittr %>%
 #' @import leaflet
 #' @import leaflet.extras